Fixed CI/CD checks for sorter_sort_data, #4
