+++ src/SortDataBase.h
@@ -4,6 +4,39 @@
 
 #include "OpState.h"
 
+#include <deque>
+#include <list>
+#include <vector>
+
+template <template <typename...> class Container>
+struct is_std_vector : std::false_type
+{
+};
+
+template <> struct is_std_vector<std::vector> : std::true_type
+{
+};
+
+
+template <template <typename...> class Container>
+struct is_std_list : std::false_type
+{
+};
+
+template <> struct is_std_list<std::list> : std::true_type
+{
+};
+
+
+template <template <typename...> class Container>
+struct is_std_deque : std::false_type
+{
+};
+
+template <> struct is_std_deque<std::deque> : std::true_type
+{
+};
+
 
 #ifdef _WIN32
 #ifdef BUILDING_CLIENT_DLL
