+++ src/SortTechFactory.h
@@ -11,7 +11,6 @@
 #include "Radix.h"
 #include "Select.h"
 
-#include <deque>
 #include <memory>
 
 enum SortTechType
@@ -23,35 +22,6 @@ enum SortTechType
     STT_QUICK,
     STT_RADIX,
     STT_SELECT
-};
-
-template <template <typename...> class Container>
-struct is_std_vector : std::false_type
-{
-};
-
-template <> struct is_std_vector<std::vector> : std::true_type
-{
-};
-
-
-template <template <typename...> class Container>
-struct is_std_list : std::false_type
-{
-};
-
-template <> struct is_std_list<std::list> : std::true_type
-{
-};
-
-
-template <template <typename...> class Container>
-struct is_std_deque : std::false_type
-{
-};
-
-template <> struct is_std_deque<std::deque> : std::true_type
-{
 };
 
 // This class extends the "factory" design pattern and is responsible for
